fix(backend): widen tokens_cached_in to BIGINT to survive Postgres
task_sessions.tokens_cached_in was declared INTEGER, which is 64-bit on
SQLite but int4 (max 2,147,483,647) on Postgres. office_cost_events.
tokens_cached_in routinely accumulates well past that ceiling for a
long-running session (98,805,109 was measured on one already-completed
task). Because IncrementTaskSessionUsage and backfillSessionTokensCachedIn
each set/recompute all four cost columns in a single UPDATE, one session
crossing the ceiling made the whole statement fail - silently, since
MigrateLogger.Apply downgrades unknown errors to a Warn - taking
tokens_in/tokens_out/cost_subcents down with it for that session, and
permanently disabling the self-healing backfill for every session on the
next boot.

Widen the column to BIGINT at all three declaration sites (fresh-install
DDL, the ADD COLUMN migration, and the legacy table-rebuild). Verified
portable to SQLite (BIGINT carries INTEGER affinity, full 64-bit range).
No repair migration is needed: the column is unreleased.

Adds session_usage_postgres_test.go, following the slack_retirement_
postgres_test.go precedent (skips unless KANDEV_TEST_POSTGRES_DSN is
set) - the SQLite-only tests could never have caught this, since
SQLite's INTEGER never overflows. Covers officeCostEventsTableExists on
both sides of table creation, the backfill sum, and specifically a
session whose ledger sums past 2,147,483,647 on a real Postgres
instance - the regression test for this bug, which fails on an INTEGER
column and passes once it is BIGINT.

+++ apps/backend/internal/task/repository/sqlite/base_migrations.go
@@ -39,7 +39,15 @@ func (r *Repository) migrateSessionsAddCostColumns() {
 	r.migrate.Apply("task_sessions.cost_subcents", `ALTER TABLE task_sessions ADD COLUMN cost_subcents INTEGER NOT NULL DEFAULT 0`)
 	r.migrate.Apply("task_sessions.tokens_in", `ALTER TABLE task_sessions ADD COLUMN tokens_in INTEGER NOT NULL DEFAULT 0`)
 	r.migrate.Apply("task_sessions.tokens_out", `ALTER TABLE task_sessions ADD COLUMN tokens_out INTEGER NOT NULL DEFAULT 0`)
-	r.migrate.Apply("task_sessions.tokens_cached_in", `ALTER TABLE task_sessions ADD COLUMN tokens_cached_in INTEGER NOT NULL DEFAULT 0`)
+	// BIGINT, not INTEGER: office_cost_events.tokens_cached_in routinely
+	// accumulates well past int4's 2,147,483,647 ceiling over a long-running
+	// session (the reported bug measured up to 98,805,109 on one already-
+	// completed task). SQLite's INTEGER is 64-bit regardless, but on Postgres
+	// INTEGER is int4 - an overflowing session would abort the single
+	// multi-column UPDATE in IncrementTaskSessionUsage and the single
+	// table-wide UPDATE in backfillSessionTokensCachedIn, silently taking
+	// tokens_in/tokens_out/cost_subcents down with it for that session.
+	r.migrate.Apply("task_sessions.tokens_cached_in", `ALTER TABLE task_sessions ADD COLUMN tokens_cached_in BIGINT NOT NULL DEFAULT 0`)
 }
 
 // officeCostEventsTableExists reports whether the office cost ledger has
@@ -707,7 +715,7 @@ func (r *Repository) migrateSessionsRemoveAgentExecutionID() error {
 			task_environment_id TEXT DEFAULT '',
 			cost_subcents INTEGER NOT NULL DEFAULT 0,
 			tokens_in INTEGER NOT NULL DEFAULT 0,
-			tokens_cached_in INTEGER NOT NULL DEFAULT 0,
+			tokens_cached_in BIGINT NOT NULL DEFAULT 0,
 			tokens_out INTEGER NOT NULL DEFAULT 0,
 			FOREIGN KEY (task_id) REFERENCES tasks(id) ON DELETE CASCADE
 		)`,
